+++ sdk/resourcemanager/powerbiembedded/armpowerbiembedded/zz_generated_workspacecollections_client.go
@@ -1,5 +1,5 @@
-//go:build go1.16
-// +build go1.16
+//go:build go1.18
+// +build go1.18
 
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
@@ -14,6 +14,7 @@ import (
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
 	armruntime "github.com/Azure/azure-sdk-for-go/sdk/azcore/arm/runtime"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/cloud"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
 	"net/http"
@@ -34,20 +35,24 @@ type WorkspaceCollectionsClient struct {
 // ID forms part of the URI for every service call.
 // credential - used to authorize requests. Usually a credential from azidentity.
 // options - pass nil to accept the default values.
-func NewWorkspaceCollectionsClient(subscriptionID string, credential azcore.TokenCredential, options *arm.ClientOptions) *WorkspaceCollectionsClient {
-	cp := arm.ClientOptions{}
-	if options != nil {
-		cp = *options
+func NewWorkspaceCollectionsClient(subscriptionID string, credential azcore.TokenCredential, options *arm.ClientOptions) (*WorkspaceCollectionsClient, error) {
+	if options == nil {
+		options = &arm.ClientOptions{}
 	}
-	if len(cp.Endpoint) == 0 {
-		cp.Endpoint = arm.AzurePublicCloud
+	ep := cloud.AzurePublicCloud.Services[cloud.ResourceManager].Endpoint
+	if c, ok := options.Cloud.Services[cloud.ResourceManager]; ok {
+		ep = c.Endpoint
+	}
+	pl, err := armruntime.NewPipeline(moduleName, moduleVersion, credential, runtime.PipelineOptions{}, options)
+	if err != nil {
+		return nil, err
 	}
 	client := &WorkspaceCollectionsClient{
 		subscriptionID: subscriptionID,
-		host:           string(cp.Endpoint),
-		pl:             armruntime.NewPipeline(moduleName, moduleVersion, credential, runtime.PipelineOptions{}, &cp),
+		host:           ep,
+		pl:             pl,
 	}
-	return client
+	return client, nil
 }
 
 // CheckNameAvailability - Verify the specified Power BI Workspace Collection name is valid and not already in use.
@@ -95,7 +100,7 @@ func (client *WorkspaceCollectionsClient) checkNameAvailabilityCreateRequest(ctx
 
 // checkNameAvailabilityHandleResponse handles the CheckNameAvailability response.
 func (client *WorkspaceCollectionsClient) checkNameAvailabilityHandleResponse(resp *http.Response) (WorkspaceCollectionsClientCheckNameAvailabilityResponse, error) {
-	result := WorkspaceCollectionsClientCheckNameAvailabilityResponse{RawResponse: resp}
+	result := WorkspaceCollectionsClientCheckNameAvailabilityResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.CheckNameResponse); err != nil {
 		return WorkspaceCollectionsClientCheckNameAvailabilityResponse{}, err
 	}
@@ -154,7 +159,7 @@ func (client *WorkspaceCollectionsClient) createCreateRequest(ctx context.Contex
 
 // createHandleResponse handles the Create response.
 func (client *WorkspaceCollectionsClient) createHandleResponse(resp *http.Response) (WorkspaceCollectionsClientCreateResponse, error) {
-	result := WorkspaceCollectionsClientCreateResponse{RawResponse: resp}
+	result := WorkspaceCollectionsClientCreateResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.WorkspaceCollection); err != nil {
 		return WorkspaceCollectionsClientCreateResponse{}, err
 	}
@@ -167,22 +172,16 @@ func (client *WorkspaceCollectionsClient) createHandleResponse(resp *http.Respon
 // workspaceCollectionName - Power BI Embedded Workspace Collection name
 // options - WorkspaceCollectionsClientBeginDeleteOptions contains the optional parameters for the WorkspaceCollectionsClient.BeginDelete
 // method.
-func (client *WorkspaceCollectionsClient) BeginDelete(ctx context.Context, resourceGroupName string, workspaceCollectionName string, options *WorkspaceCollectionsClientBeginDeleteOptions) (WorkspaceCollectionsClientDeletePollerResponse, error) {
-	resp, err := client.deleteOperation(ctx, resourceGroupName, workspaceCollectionName, options)
-	if err != nil {
-		return WorkspaceCollectionsClientDeletePollerResponse{}, err
+func (client *WorkspaceCollectionsClient) BeginDelete(ctx context.Context, resourceGroupName string, workspaceCollectionName string, options *WorkspaceCollectionsClientBeginDeleteOptions) (*armruntime.Poller[WorkspaceCollectionsClientDeleteResponse], error) {
+	if options == nil || options.ResumeToken == "" {
+		resp, err := client.deleteOperation(ctx, resourceGroupName, workspaceCollectionName, options)
+		if err != nil {
+			return nil, err
+		}
+		return armruntime.NewPoller[WorkspaceCollectionsClientDeleteResponse](resp, client.pl, nil)
+	} else {
+		return armruntime.NewPollerFromResumeToken[WorkspaceCollectionsClientDeleteResponse](options.ResumeToken, client.pl, nil)
 	}
-	result := WorkspaceCollectionsClientDeletePollerResponse{
-		RawResponse: resp,
-	}
-	pt, err := armruntime.NewPoller("WorkspaceCollectionsClient.Delete", "", resp, client.pl)
-	if err != nil {
-		return WorkspaceCollectionsClientDeletePollerResponse{}, err
-	}
-	result.Poller = &WorkspaceCollectionsClientDeletePoller{
-		pt: pt,
-	}
-	return result, nil
 }
 
 // Delete - Delete a Power BI Workspace Collection.
@@ -277,7 +276,7 @@ func (client *WorkspaceCollectionsClient) getAccessKeysCreateRequest(ctx context
 
 // getAccessKeysHandleResponse handles the GetAccessKeys response.
 func (client *WorkspaceCollectionsClient) getAccessKeysHandleResponse(resp *http.Response) (WorkspaceCollectionsClientGetAccessKeysResponse, error) {
-	result := WorkspaceCollectionsClientGetAccessKeysResponse{RawResponse: resp}
+	result := WorkspaceCollectionsClientGetAccessKeysResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.WorkspaceCollectionAccessKeys); err != nil {
 		return WorkspaceCollectionsClientGetAccessKeysResponse{}, err
 	}
@@ -333,7 +332,7 @@ func (client *WorkspaceCollectionsClient) getByNameCreateRequest(ctx context.Con
 
 // getByNameHandleResponse handles the GetByName response.
 func (client *WorkspaceCollectionsClient) getByNameHandleResponse(resp *http.Response) (WorkspaceCollectionsClientGetByNameResponse, error) {
-	result := WorkspaceCollectionsClientGetByNameResponse{RawResponse: resp}
+	result := WorkspaceCollectionsClientGetByNameResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.WorkspaceCollection); err != nil {
 		return WorkspaceCollectionsClientGetByNameResponse{}, err
 	}
@@ -345,19 +344,26 @@ func (client *WorkspaceCollectionsClient) getByNameHandleResponse(resp *http.Res
 // resourceGroupName - Azure resource group
 // options - WorkspaceCollectionsClientListByResourceGroupOptions contains the optional parameters for the WorkspaceCollectionsClient.ListByResourceGroup
 // method.
-func (client *WorkspaceCollectionsClient) ListByResourceGroup(ctx context.Context, resourceGroupName string, options *WorkspaceCollectionsClientListByResourceGroupOptions) (WorkspaceCollectionsClientListByResourceGroupResponse, error) {
-	req, err := client.listByResourceGroupCreateRequest(ctx, resourceGroupName, options)
-	if err != nil {
-		return WorkspaceCollectionsClientListByResourceGroupResponse{}, err
-	}
-	resp, err := client.pl.Do(req)
-	if err != nil {
-		return WorkspaceCollectionsClientListByResourceGroupResponse{}, err
-	}
-	if !runtime.HasStatusCode(resp, http.StatusOK) {
-		return WorkspaceCollectionsClientListByResourceGroupResponse{}, runtime.NewResponseError(resp)
-	}
-	return client.listByResourceGroupHandleResponse(resp)
+func (client *WorkspaceCollectionsClient) ListByResourceGroup(resourceGroupName string, options *WorkspaceCollectionsClientListByResourceGroupOptions) *runtime.Pager[WorkspaceCollectionsClientListByResourceGroupResponse] {
+	return runtime.NewPager(runtime.PageProcessor[WorkspaceCollectionsClientListByResourceGroupResponse]{
+		More: func(page WorkspaceCollectionsClientListByResourceGroupResponse) bool {
+			return false
+		},
+		Fetcher: func(ctx context.Context, page *WorkspaceCollectionsClientListByResourceGroupResponse) (WorkspaceCollectionsClientListByResourceGroupResponse, error) {
+			req, err := client.listByResourceGroupCreateRequest(ctx, resourceGroupName, options)
+			if err != nil {
+				return WorkspaceCollectionsClientListByResourceGroupResponse{}, err
+			}
+			resp, err := client.pl.Do(req)
+			if err != nil {
+				return WorkspaceCollectionsClientListByResourceGroupResponse{}, err
+			}
+			if !runtime.HasStatusCode(resp, http.StatusOK) {
+				return WorkspaceCollectionsClientListByResourceGroupResponse{}, runtime.NewResponseError(resp)
+			}
+			return client.listByResourceGroupHandleResponse(resp)
+		},
+	})
 }
 
 // listByResourceGroupCreateRequest creates the ListByResourceGroup request.
@@ -384,7 +390,7 @@ func (client *WorkspaceCollectionsClient) listByResourceGroupCreateRequest(ctx c
 
 // listByResourceGroupHandleResponse handles the ListByResourceGroup response.
 func (client *WorkspaceCollectionsClient) listByResourceGroupHandleResponse(resp *http.Response) (WorkspaceCollectionsClientListByResourceGroupResponse, error) {
-	result := WorkspaceCollectionsClientListByResourceGroupResponse{RawResponse: resp}
+	result := WorkspaceCollectionsClientListByResourceGroupResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.WorkspaceCollectionList); err != nil {
 		return WorkspaceCollectionsClientListByResourceGroupResponse{}, err
 	}
@@ -395,19 +401,26 @@ func (client *WorkspaceCollectionsClient) listByResourceGroupHandleResponse(resp
 // If the operation fails it returns an *azcore.ResponseError type.
 // options - WorkspaceCollectionsClientListBySubscriptionOptions contains the optional parameters for the WorkspaceCollectionsClient.ListBySubscription
 // method.
-func (client *WorkspaceCollectionsClient) ListBySubscription(ctx context.Context, options *WorkspaceCollectionsClientListBySubscriptionOptions) (WorkspaceCollectionsClientListBySubscriptionResponse, error) {
-	req, err := client.listBySubscriptionCreateRequest(ctx, options)
-	if err != nil {
-		return WorkspaceCollectionsClientListBySubscriptionResponse{}, err
-	}
-	resp, err := client.pl.Do(req)
-	if err != nil {
-		return WorkspaceCollectionsClientListBySubscriptionResponse{}, err
-	}
-	if !runtime.HasStatusCode(resp, http.StatusOK) {
-		return WorkspaceCollectionsClientListBySubscriptionResponse{}, runtime.NewResponseError(resp)
-	}
-	return client.listBySubscriptionHandleResponse(resp)
+func (client *WorkspaceCollectionsClient) ListBySubscription(options *WorkspaceCollectionsClientListBySubscriptionOptions) *runtime.Pager[WorkspaceCollectionsClientListBySubscriptionResponse] {
+	return runtime.NewPager(runtime.PageProcessor[WorkspaceCollectionsClientListBySubscriptionResponse]{
+		More: func(page WorkspaceCollectionsClientListBySubscriptionResponse) bool {
+			return false
+		},
+		Fetcher: func(ctx context.Context, page *WorkspaceCollectionsClientListBySubscriptionResponse) (WorkspaceCollectionsClientListBySubscriptionResponse, error) {
+			req, err := client.listBySubscriptionCreateRequest(ctx, options)
+			if err != nil {
+				return WorkspaceCollectionsClientListBySubscriptionResponse{}, err
+			}
+			resp, err := client.pl.Do(req)
+			if err != nil {
+				return WorkspaceCollectionsClientListBySubscriptionResponse{}, err
+			}
+			if !runtime.HasStatusCode(resp, http.StatusOK) {
+				return WorkspaceCollectionsClientListBySubscriptionResponse{}, runtime.NewResponseError(resp)
+			}
+			return client.listBySubscriptionHandleResponse(resp)
+		},
+	})
 }
 
 // listBySubscriptionCreateRequest creates the ListBySubscription request.
@@ -430,7 +443,7 @@ func (client *WorkspaceCollectionsClient) listBySubscriptionCreateRequest(ctx co
 
 // listBySubscriptionHandleResponse handles the ListBySubscription response.
 func (client *WorkspaceCollectionsClient) listBySubscriptionHandleResponse(resp *http.Response) (WorkspaceCollectionsClientListBySubscriptionResponse, error) {
-	result := WorkspaceCollectionsClientListBySubscriptionResponse{RawResponse: resp}
+	result := WorkspaceCollectionsClientListBySubscriptionResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.WorkspaceCollectionList); err != nil {
 		return WorkspaceCollectionsClientListBySubscriptionResponse{}, err
 	}
@@ -455,7 +468,7 @@ func (client *WorkspaceCollectionsClient) Migrate(ctx context.Context, resourceG
 	if !runtime.HasStatusCode(resp, http.StatusOK) {
 		return WorkspaceCollectionsClientMigrateResponse{}, runtime.NewResponseError(resp)
 	}
-	return WorkspaceCollectionsClientMigrateResponse{RawResponse: resp}, nil
+	return WorkspaceCollectionsClientMigrateResponse{}, nil
 }
 
 // migrateCreateRequest creates the Migrate request.
@@ -530,7 +543,7 @@ func (client *WorkspaceCollectionsClient) regenerateKeyCreateRequest(ctx context
 
 // regenerateKeyHandleResponse handles the RegenerateKey response.
 func (client *WorkspaceCollectionsClient) regenerateKeyHandleResponse(resp *http.Response) (WorkspaceCollectionsClientRegenerateKeyResponse, error) {
-	result := WorkspaceCollectionsClientRegenerateKeyResponse{RawResponse: resp}
+	result := WorkspaceCollectionsClientRegenerateKeyResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.WorkspaceCollectionAccessKeys); err != nil {
 		return WorkspaceCollectionsClientRegenerateKeyResponse{}, err
 	}
@@ -587,7 +600,7 @@ func (client *WorkspaceCollectionsClient) updateCreateRequest(ctx context.Contex
 
 // updateHandleResponse handles the Update response.
 func (client *WorkspaceCollectionsClient) updateHandleResponse(resp *http.Response) (WorkspaceCollectionsClientUpdateResponse, error) {
-	result := WorkspaceCollectionsClientUpdateResponse{RawResponse: resp}
+	result := WorkspaceCollectionsClientUpdateResponse{}
 	if err := runtime.UnmarshalAsJSON(resp, &result.WorkspaceCollection); err != nil {
 		return WorkspaceCollectionsClientUpdateResponse{}, err
 	}